requirements.txt
discord.py
shortuuid
python-dotenv
firebase_admin
sendgrid